Fixed AmbientLightRooms again, don't revert
No offense, but the changes made to AmbientLightRooms were completely incorrect. The sole purpose of the function is to add some ambient lighting to rooms when the NVG are on, which is why they need additive blending.

The "isAvailRT" code is faulty, so I decided to just shift the texture layers by one and put AmbientLightRoomTex on layer 0, which is how it's supposed to be applied.

The problem with the brightness was that 100 is too much, apparently. It looks good at 20, so I just set it to that value. I also properly fixed vertex-lit models in LoadRMesh.