Removed "bb_fmod.dll" and only use "fmod.dll" for sound streaming (StreamSound_Strict function was removed, details are written in this commit here) | added a "GlobalMemoryStatus" system (altough it needs testing to see if it's actually useful) | hopefully fixed the issue with the player being able not to trigger the 860-2 alert music
Here's the explanation why "StreamSound_Strict" was removed: I tried multiple times to rewrite the function, so that "FSOUND_Stream_Open" stuff and such doesn't need to be copied over and over again. Altough whenever I added something in a seperate loading function (like StreamSound_Strict), then somethign wouldn't work correctly, like the volume adjustment, the pausing and continuing and at some occasions the sound channels returned -1 when trying to play the stream (this means that an error occured). Actually, whenever I loaded a stream using a function like StreamSound_Strict, then it returned an Int number, but I think the fmod channels didn't recognized it as I think another data type needs to be returned for this that Blitz3D doesn't really understand, but I might be wrong. In the way I did it it works just fine, altough the mechanics should still be tested (especially in GateB) to see if there are still issues occuring.