Replaced DirectMovie commands with another library (BlitzMovie).
This will hopefully fix the crashes that some users had relating to the incompability of DirectMovie.It's also worth to mention that the audio is seperated from the video. There are 2 reasons for this: The first reason is that some people had problems with the audio either during or after the video or general audio problems happened during the playing of the avi file. The other reason is because the BlitzMovie plugin doesn't has a command to check if the movie is finished playing (it technically has, but it either is for a different purpose or it is broken), so the audio file is determining when the program can continue without the user having any need to press a button to continue.