Fix blue NVG sometimes showing 2 digits past decimal point instead of 1
f2s is a function that takes a float and returns a string containing the entire integer part of the float followed by the specified amount of digits (count) of the fractional part. Len(Int(n)) is supposed to represent the length of the integer part but this has rounding issues (if n is 9.8, Len(Int(n)) returns 2, because Int(n) returns 10). Using Int(Str(n)) instead ensures that the integer part is not rounded.