Fix log scaling
Because the creation of the log items happened before the downscaling of the parent forest tile, the logs were being downscaled along with the tile. This issue was already countered by the fact that the logs' item templates had larger scale values, but because the downscaling of the parent forest tile only happens when creating a new game, logs would be large when playing on a loaded save. The solution is to give the logs' item templates their true scale values, and only create a log after the downscaling of its parent forest tile.