Fix DrawTiledImageRect
DrawTiledImageRect takes a patch from an image and tiles a rectangle with copies of that patch. It starts in the top left corner and moves down. When a patch eventually surpasses the bottom of the rectangle, the patch is snipped to fit, then the function moves to the next column. The problem is that the original patch is what got snipped, not the copy. This means all future copies of the patch are shorter, and continue to get shorter the further to the right the function progresses, resulting in a very strange image. The problem is more apparent the longer the rectangle is.

The solution is to only snip the copy by passing the adjusted srcheight value to DrawImageRect directly rather than modifying srcheight. It's okay to modify srcwidth as we won't need any more columns after reaching the right end of the rectangle. We only need to modify it once though, so the line has been moved outside the y-loop. Some expressions have also been simplified.